+++ changelog.md
@@ -1,10 +1,16 @@
 ## [0.14.0]
 
 ### Breaking Changes
-* None.
+* Removed `GrpcChannelFactory`. The connect methods in the `Connect` object should be used instead. `GrpcChannelBuilder` may be used if more customisation is
+  required (e.g. TLS client authentication).
 
 ### New Features
-* None.
+* Added methods for connecting to the gRPC service in a utility object named `Connect`:
+  * `Connect.connectInsecure`: Used to connect to a gRPC service that does not use SSL/TLS.
+  * `Connect.connectTls`: Used to connect to a gRPC service that uses SSL/TLS, without user authentication.
+  * `Connect.connectWithSecret`: Used to connect to a gRPC service that uses SSL/TLS, using the OAuth 2.0 client credential flow.
+  * `Connect.connectWithPassword`: Used to connect to a gRPC service that uses SSL/TLS, using the OAuth 2.0 password grant flow.
+
 
 ### Enhancements
 * None.